+++ programs/proof-api/Cargo.toml
@@ -8,7 +8,7 @@ license    = { workspace = true }
 [dependencies]
 proof-api-core                 = { workspace = true, default-features = false }
 proof-api-eth-to-cosmos        = { workspace = true, default-features = false }
-proof-api-eth-to-cosmos-compat = { workspace = true, default-features = false }
+proof-api-eth-to-cosmos-v1-3-0 = { workspace = true, default-features = false }
 proof-api-cosmos-to-eth        = { workspace = true, default-features = false }
 proof-api-eth-to-eth           = { workspace = true, default-features = false }
 proof-api-cosmos-to-cosmos     = { workspace = true, default-features = false }
